pytest: Separate tracker tests
